+++ src/zoid/saved-payment-methods/component.jsx
@@ -578,6 +578,7 @@ export const getSavedPaymentMethodsComponent: () => SavedPaymentMethodsComponent
         partnerAttributionID: {
           type: "string",
           required: false,
+          queryParam: true,
           value: getPartnerAttributionID,
         },